refactor: renamed useInfiniteAvailableAgentsList to useAvailableAgentsList
