+++ apps/meteor/client/components/AutoCompleteAgentWithoutExtension.tsx
@@ -2,7 +2,7 @@ import { PaginatedSelectFiltered } from '@rocket.chat/fuselage';
 import { useDebouncedValue } from '@rocket.chat/fuselage-hooks';
 import { memo, useState } from 'react';
 
-import { useInfiniteAvailableAgentsList } from './Omnichannel/hooks/useInfiniteAvailableAgentsList';
+import { useAvailableAgentsList } from './Omnichannel/hooks/useAvailableAgentsList';
 
 type AutoCompleteAgentProps = {
 	value: string;
@@ -16,7 +16,7 @@ const AutoCompleteAgentWithoutExtension = ({ value, currentExtension, onChange, 
 
 	const debouncedAgentsFilter = useDebouncedValue(agentsFilter as string, 500);
 
-	const { data: agentsItems, fetchNextPage } = useInfiniteAvailableAgentsList({
+	const { data: agentsItems, fetchNextPage } = useAvailableAgentsList({
 		text: debouncedAgentsFilter,
 		includeExtension: currentExtension,
 	});
